feat: new function system
